requirements.txt
fastapi
uvicorn

pydantic

pandas
boto3

python-dotenv
